Preparing v3.8 ...
[Flyleaf]
* Decoder: Will not provide codec during avcodec_open2 (only once during avcodec_alloc_context3) as it could cause issues
* VideoDecoder: Will not set pix_fmt to D3D11 and uses extra_hw_frames instead of initial_pool_size (could cause issues eg. with AV1)
* VideoDecoder: Handles another possible VA failure and uses SWFallback
* VideoStream: Should fix an issue with invalid rotation

[Dependencies]
* Replaces FFmpeg.Autogen with Flyleaf.FFmpeg.Bindings
* Replaces Microsoft ILoggger with a generic custom callback
* Updates FFmpeg libs to v7.1 (patched *and mingw downgrade)
* Updates Vortice libs to v3.6.2

[Breaking Changes]
* Deprecates .NET Framework and .NET 6 (Support only .NET 8)
* FFmpegLogLevel changed to Flyleaf.FFmpeg.LogLevel

[Solution]
* Removes Tests/WpfNetFramework project
* Codec clean up (removes #if for other .NET versions)

[TBR]
* Audio.Filters: Cannot use min/max samples on buffer sink context
* VideoDecoder: Keyframes to be removed and focus only on packet keys

+++ FlyleafLib/MediaPlayer/Audio.cs
@@ -217,7 +217,7 @@ public class Audio : NotifyPropertyChanged
                 }
 
                 sourceVoice = xaudio2.CreateSourceVoice(waveFormat, false);
-                sourceVoice.SetSourceSampleRate(SampleRate);
+                sourceVoice.SetSourceSampleRate((uint)SampleRate);
                 sourceVoice.Start();
 
                 submittedSamples        = 0;
@@ -266,7 +266,7 @@ public class Audio : NotifyPropertyChanged
                 SamplesAdded?.Invoke(this, aFrame);
             
                 audioBuffer.AudioDataPointer= aFrame.dataPtr;
-                audioBuffer.AudioBytes      = aFrame.dataLen;
+                audioBuffer.AudioBytes      = (uint)aFrame.dataLen;
                 sourceVoice.SubmitSourceBuffer(audioBuffer);
             }
             catch (Exception e) // Happens on audio device changed/removed
